more ATA PIO stuff

+++ src/bootloader/stage2/arch/x86/io.h
@@ -1,7 +1,7 @@
 #pragma once
 
-extern "C" uint8_t inb(uint8_t port);
-extern "C" void outb(uint8_t port, uint8_t data);
+extern "C" uint8_t inb(uint16_t port);
+extern "C" void outb(uint16_t port, uint8_t data);
 
-extern "C" uint16_t inw(uint8_t port);
-extern "C" void outw(uint8_t port, uint16_t data);
+extern "C" uint16_t inw(uint16_t port);
+extern "C" void outw(uint16_t port, uint16_t data);
